sunsynk-dashboard/backend/requirements.txt
# FastAPI Backend Requirements for Sunsynk Solar Dashboard
# Phase 6: Advanced Analytics & ML-powered Optimization

# Core API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# WebSocket Support
websockets==12.0

# Authentication & Security
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database & Data Processing
influxdb-client==1.38.0
aiofiles==23.2.1

# HTTP Client (for external APIs)
aiohttp==3.9.0

# Environment & Configuration
python-dotenv==1.0.0

# Data Validation & Serialization
pydantic==2.5.0

# Phase 6: Machine Learning & Analytics
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
scipy==1.11.4

# Logging & Monitoring
structlog==23.2.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Production Server
gunicorn==21.2.0
